[server] server-for-lambda のRESTちっくなI/F仕様に追従
